ckeditor_sanpham

+++ WebApplication/Controllers/NhomMatHangController.cs
@@ -10,7 +10,7 @@ namespace WebApplication.Controllers
 {
     public class NhomMatHangController : Controller
     {
-        DatabaseEntities db = new DatabaseEntities();
+        DOANEntities db = new DOANEntities();
         public ActionResult Index()
         {
             List<DNHOMMATHANG> lst = db.DNHOMMATHANGs.OrderBy(x => x.CODE).ToList();
@@ -29,24 +29,11 @@ namespace WebApplication.Controllers
         }
 
         [HttpPost]
-        public ActionResult AddOrEdit(string id, DMATHANG temp)
+        public ActionResult AddOrEdit(string id, DNHOMMATHANG temp)
         {
             string error = "";
             try
             {
-                DNHOMMATHANG model = db.DNHOMMATHANGs.Where(x => x.ID == id).FirstOrDefault();
-                if (model == null)
-                {
-                    model = new DNHOMMATHANG();
-                    model.ID = Guid.NewGuid().ToString();
-                }
-                model.CODE = temp.CODE.Trim();
-                model.NAME = temp.NAME.Trim();
-                if (id == null) db.DNHOMMATHANGs.Add(model);
-                else
-                {
-                    db.Entry(model).State = EntityState.Modified;
-                }
                 //kiểm tra trống
                 if (temp.CODE == null || temp.CODE.Length == 0) error = "Mã nhóm hàng không được trống!";
                 else if (temp.NAME == null || temp.NAME.Length == 0) error = "Tên nhóm hàng không được trống!";
@@ -57,7 +44,26 @@ namespace WebApplication.Controllers
                     error = "Mã nhóm mặt hàng không được chứa nhóm mặt hàng khác";
                 }
 
-                if (error.Length == 0) db.SaveChanges();
+                if (error.Length == 0)
+                {
+                    DNHOMMATHANG model = db.DNHOMMATHANGs.Where(x => x.ID == id).FirstOrDefault();
+                    if (model == null)
+                    {
+                        model = new DNHOMMATHANG();
+                    }
+                    model.CODE = temp.CODE.Trim();
+                    model.NAME = temp.NAME.Trim();
+                    if (id == null)
+                    {
+                        model.ID = Guid.NewGuid().ToString();
+                        db.DNHOMMATHANGs.Add(model);
+                    }
+                    else
+                    {
+                        db.Entry(model).State = EntityState.Modified;
+                    }
+                    db.SaveChanges();
+                }
             }
             catch (Exception ex)
             {
@@ -76,8 +82,15 @@ namespace WebApplication.Controllers
                 if (entry == null) error = "Nhóm hàng không tồn tại trong hệ thống!";
                 else
                 {
-                    db.DNHOMMATHANGs.Remove(entry);
-                    db.SaveChanges();
+                    if (entry.DMATHANGs.Count > 0)
+                    {
+                        error = "có " + entry.DMATHANGs.Count + " liên kết tới nhóm hàng, không thể xóa";
+                    }
+                    else
+                    {
+                        db.DNHOMMATHANGs.Remove(entry);
+                        db.SaveChanges();
+                    }
                 }
             }
             catch (Exception ex)
